지난 자료/django_with_vue/articleback/requirements.txt
appdirs==1.4.4
attrs==19.3.0
black==19.10b0
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
defusedxml==0.6.0
Django==2.1.15
django-allauth==0.42.0
django-cors-headers==3.3.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
idna==2.9
oauthlib==3.1.0
pathspec==0.8.0
python3-openid==3.1.0
pytz==2020.1
regex==2020.6.8
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
